fix: rename Aws configuration file
