+++ packages/common/src/interfaces/IAwsConfig.ts
